pyproject.toml
[project]
name = "health-coach"
version = "0.1.0"
description = "Data-driven health and fitness tracking framework"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary",
    "pyyaml",
    "python-dotenv",
    "fastmcp",
    "garminconnect",
    "chainlit",
    "openai",
    "plotly",
    "PyJWT[crypto]",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.setuptools.packages.find]
where = ["scripts"]
include = []
